requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-decouple==3.8
PyJWT==2.6.0
Pillow==9.5.0
psycopg2-binary==2.9.6
drf-yasg==1.21.5
celery==5.2.7
redis==4.5.5
requests==2.31.0
django-filter==23.1
daphne==4.2.1
djangorestframework-simplejwt==5.5.1
gunicorn==21.2.0
whitenoise==6.4.0
